requirements.txt
# Core Deep Learning
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1

# Transformers & Vision Models
transformers>=4.30.0
timm>=0.9.0

# Image Processing & Data
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# Machine Learning & Metrics
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization & Plotting
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0

# XAI & Explainability
lime>=0.2.0
shap>=0.41.0

# Optional: For distributed training
# horovod>=0.26.0

# Development
jupyter>=1.0.0
ipython>=8.0.0
